refactor(interpreter): split comprehension formula and match tests
