+++ src/interpreter/src/expressions.rs
@@ -338,27 +338,6 @@ impl MechFunctionImpl for ValueMatrixComprehension {
     }
 }
 
-#[cfg(all(test, feature = "matrix_comprehensions", feature = "functions"))]
-mod matrix_comprehension_transaction_tests {
-    use super::*;
-
-    #[test]
-    fn transaction_state_retains_matrix_comprehension_outer_output_ref() {
-        let out = Ref::new(Value::Empty);
-        let function = ValueMatrixComprehension {
-            arguments: Vec::new(),
-            out: out.clone(),
-        };
-
-        let values = function.transaction_state_values().unwrap();
-        assert_eq!(values.len(), 1);
-        match &values[0] {
-            Value::MutableReference(root) => assert_eq!(root.addr(), out.addr()),
-            other => panic!("expected mutable-reference transaction root, got {other:?}"),
-        }
-    }
-}
-
 #[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
 impl MechFunctionFactory for ValueMatrixComprehension {
     fn new(args: FunctionArgs) -> MResult<Box<dyn MechFunction>> {
